fix(e2e): stop full-mode @agent specs cascading on shared host state
The full e2e suite failed 19 @agent specs locally. Three root causes,
all in the e2e harness (no product code changed):

- jbcentral spec poisoned the shared model registry. On a machine authed
  through the jbcentral proxy, the seeded models.json IS the auth (anthropic/
  openai baseUrl+apiKey wired to the proxy). The JetBrains AI connect/disconnect
  spec strips those fields and leaves the host disconnected, and resetState
  never restored models.json — so every @agent spec after it found an empty
  registry. Fix: snapshot the seeded models.json in globalSetup and re-seed it
  per test in resetState (drop the wire's .bak); the next page load's
  provider.status refreshes the registry.

- error-turn regressed from #61. Host-side model validation now rejects a bad
  model at session.create (before a chat exists), so the failure surfaces as a
  "Couldn't start the chat" toast, not an in-chat error turn. Assert the toast
  (carrying the rejected model id) + no chat tab + no false Done.

- setting-up-a-project could delete the shared fixture repo. It drives a real
  agent with bash in a worktree of the shared repo; a stray destructive command
  removed it, cascading into every later test's reset. Fix: extract fixture-repo
  seeding into fixtures/repo.ts and self-heal it in resetState, bounding the
  blast radius to the one spec.

Full suite: 19 systemic failures -> 0 (one remaining miss is @agent
non-determinism — visualize's optional `recommended` flag — which passes in
isolation and is absorbed by CI retries).
