+++ e2e/global-setup.ts
@@ -1,4 +1,3 @@
-import { execFileSync } from "node:child_process";
 import { copyFileSync, existsSync, mkdirSync, rmSync, writeFileSync } from "node:fs";
 import { homedir } from "node:os";
 import { join } from "node:path";
@@ -6,13 +5,15 @@ import {
 	E2E_DATA_DIR,
 	E2E_FIXTURE_REPO,
 	E2E_PI_AGENT_DIR,
+	E2E_PI_MODELS_SEED,
 	E2E_PICK_DIR_POINTER,
 } from "./fixtures/paths";
+import { seedFixtureRepo } from "./fixtures/repo";
 
 /** Fresh, isolated state dir + a throwaway git repo to open as a project. (Runs under node, not bun.) */
 export default function globalSetup(): void {
 	rmSync(E2E_DATA_DIR, { recursive: true, force: true });
-	mkdirSync(E2E_FIXTURE_REPO, { recursive: true });
+	mkdirSync(E2E_DATA_DIR, { recursive: true });
 
 	// Isolated pi agent dir: copy the user's provider/auth config so a real provider works (the `@agent`
 	// suite needs it — auth lives across BOTH `auth.json` (OAuth providers) and `models.json` (providers
@@ -27,6 +28,12 @@ export default function globalSetup(): void {
 		const src = join(userAgentDir, file);
 		if (existsSync(src)) copyFileSync(src, join(E2E_PI_AGENT_DIR, file));
 	}
+	// Keep a pristine snapshot of the seeded models.json: the JetBrains AI spec mutates the shared agent-dir
+	// copy (proxy wire/unwire), so `resetState` restores it per test (see E2E_PI_MODELS_SEED). No file means
+	// the dev authed via auth.json only — reset then just clears any test-written models.json.
+	const modelsSeedSrc = join(userAgentDir, "models.json");
+	if (existsSync(modelsSeedSrc)) copyFileSync(modelsSeedSrc, E2E_PI_MODELS_SEED);
+	else rmSync(E2E_PI_MODELS_SEED, { force: true });
 	const [provider, ...idParts] = (
 		process.env.THINKRAIL_E2E_MODEL ?? "anthropic/claude-opus-4-8"
 	).split("/");
@@ -35,82 +42,9 @@ export default function globalSetup(): void {
 		`${JSON.stringify({ defaultProvider: provider, defaultModel: idParts.join("/"), defaultThinkingLevel: "low" }, null, 2)}\n`,
 	);
 
-	const git = (...args: string[]) =>
-		execFileSync("git", ["-C", E2E_FIXTURE_REPO, ...args], { stdio: "ignore" });
-	git("init", "-b", "main");
-	git("config", "user.email", "e2e@thinkrail.test");
-	git("config", "user.name", "ThinkRail E2E");
-	writeFileSync(join(E2E_FIXTURE_REPO, "README.md"), "# sample-project\n");
-	// A non-markdown file so the editor suite can assert the source-only path (no rendered-view toggle).
-	writeFileSync(join(E2E_FIXTURE_REPO, "notes.txt"), "plain-text-fixture\n");
-	// A markdown file exercising GitHub-style alert callouts (+ a plain blockquote for contrast), for the
-	// rendered-preview suite (see e2e/markdown-alerts.spec.ts).
-	writeFileSync(
-		join(E2E_FIXTURE_REPO, "ALERTS.md"),
-		[
-			"# Alert callouts",
-			"",
-			"> [!NOTE]",
-			"> Useful information users should know.",
-			"",
-			"> [!TIP]",
-			"> Helpful advice for doing things better.",
-			"",
-			"> [!IMPORTANT]",
-			"> Key information to achieve a goal.",
-			"",
-			"> [!WARNING]",
-			"> Urgent info needing immediate attention.",
-			"",
-			"> [!CAUTION]",
-			"> Advises about risky outcomes.",
-			"",
-			"> A plain blockquote, no marker, so it stays a quote.",
-			"",
-		].join("\n"),
-	);
-	// A doc + image for the rendered-preview link/anchor/image suite (see e2e/markdown-links.spec.ts):
-	// a relative file link (opens the target tab), an in-doc anchor, and a relative image (host /files route).
-	writeFileSync(
-		join(E2E_FIXTURE_REPO, "LINKS.md"),
-		[
-			"# Link demo",
-			"",
-			"Jump to [Section two](#section-two), open [the spec](SPEC.md), and see the logo:",
-			"",
-			"![logo](logo.png)",
-			"",
-			"## Section two",
-			"",
-			"Target of the in-document anchor.",
-			"",
-		].join("\n"),
-	);
-	// A real 1x1 PNG so the relative-image path serves actual image bytes over the host /files route.
-	writeFileSync(
-		join(E2E_FIXTURE_REPO, "logo.png"),
-		Buffer.from(
-			"iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNk+M8AAAMCAoGB9x0AAAAASUVORK5CYII=",
-			"base64",
-		),
-	);
-	// A seed spec so the @agent `spec-tools` suite has a deterministic `spec_grep` match, proving the
-	// bundled `pi-spec-graph` extension is wired into a live session (see e2e/spec-tools.live.spec.ts). The
-	// token SPECGRAPHPROBE is distinctive so a match can't be an echo of the query; the file path it lives
-	// in is the proof of a real hit.
-	writeFileSync(
-		join(E2E_FIXTURE_REPO, "SPEC.md"),
-		"---\nid: sample-root\ntype: goal-and-requirements\ntitle: Sample Project\n---\n\n## Goal\n\nA throwaway fixture project for the thinkrail e2e suite. It carries the token SPECGRAPHPROBE so spec_grep has a deterministic match to find.\n",
-	);
-	// A child spec under the root, so the Specs viewer has a deterministic parent-tree to render
-	// (see e2e/specs-panel.spec.ts).
-	mkdirSync(join(E2E_FIXTURE_REPO, "module-a"), { recursive: true });
-	writeFileSync(
-		join(E2E_FIXTURE_REPO, "module-a", "SPEC.md"),
-		"---\nid: sample-module\ntype: module-design\nstatus: active\ntitle: Sample Module\nparent: sample-root\n---\n\n## Responsibility\n\nA fixture module spec, child of sample-root.\n",
-	);
-	git("add", "-A");
-	git("commit", "-m", "init");
+	// Seed the shared fixture repo (git init + seed files + commit). Shared with per-test `resetState`, which
+	// re-seeds it if a flaky @agent spec damages the repo (see fixtures/repo.ts).
+	seedFixtureRepo();
 
 	// Point the stubbed picker (its `THINKRAIL_PICK_DIR` names this file) at the git fixture by default;
 	// a test can rewrite it to hand the picker a different folder without restarting the shared host.
